requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
torchaudio
ultralytics  # YOLOv8
python-multipart  # для загрузки изображений
pydantic
typing